Safari-609.1.11.1/LayoutTests/platform/win/platform/mac/fast/forms/input-appearance-spinbutton-size-expected.txt

Test sizes of outer-spin-button. A spin-button height should be equal to or less than the corresponding text field height.

PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS spinButton.offsetHeight <= input.offsetHeight + 6 is true
PASS successfullyParsed is true

TEST COMPLETE
